requirements.txt
Django==3.2.12
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.11.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.7
